Move alerts into HomeVC extension
